Validaciones
Elimino métodos de validación de ID que no usan más.
Incorporo validación del CUIT según el formato XX-YYYYYYYY-Z
Modifico la variable (en C# y en la db) para que soporte números
grandes.

+++ Base de datos/Consulta para crear y llenar base de datos.sql
@@ -24,7 +24,7 @@ id int NOT NULL IDENTITY(1,1) PRIMARY KEY,
 descripcion varchar(255),
 capacidad int,
 destino int FOREIGN KEY REFERENCES Destino(id),
-cuit int NOT NULL UNIQUE,
+cuit bigint NOT NULL UNIQUE,
 aceptaMascota bit,
 inicioActividad date,
 eliminado bit);
